Rename components from tree to graph
