+++ utils/client/dependencyGraphParser.ts
